% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{argentum_cache_clear}
\alias{argentum_cache_clear}
\title{Clear the argentum cache}
\usage{
argentum_cache_clear(what = c("all", "memory", "disk"))
}
\arguments{
\item{what}{Which tier to clear: \code{"all"}, \code{"memory"} or \code{"disk"}.}
}
\value{
Invisibly, the number of on-disk files removed.
}
\description{
Clear the argentum cache
}
\examples{
argentum_cache_clear("memory")
}
