% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_capabilities.R
\name{argentum_get_capabilities}
\alias{argentum_get_capabilities}
\title{Get WMS or WFS Capabilities}
\usage{
argentum_get_capabilities(url, max_tries = 3, timeout = 30)
}
\arguments{
\item{url}{The full URL of the WMS or WFS service, including all parameters}

\item{max_tries}{The maximum number of attempts to connect}

\item{timeout}{The timeout for each attempt in seconds}
}
\value{
An XML document containing the service capabilities
}
\description{
Get WMS or WFS Capabilities
}
\examples{
\donttest{
try({
  capabilities <- argentum_get_capabilities("http://example.com/wfs")
  print(capabilities)
})
}
}
