% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_import.R
\name{argentum_interactive_import}
\alias{argentum_interactive_import}
\title{Import WFS Layer Interactively}
\usage{
argentum_interactive_import()
}
\value{
An sf object
}
\description{
Import WFS Layer Interactively
}
\examples{
\donttest{
sf_layer <- argentum_interactive_import()
}
}
