% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_organizations.R
\name{argentum_list_organizations}
\alias{argentum_list_organizations}
\title{List Organizations with WMS and WFS Services}
\usage{
argentum_list_organizations()
}
\value{
A data frame containing organization information including WMS and WFS URLs
}
\description{
List Organizations with WMS and WFS Services
}
