% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_url.R
\name{clean_url}
\alias{clean_url}
\title{Clean and format URL}
\usage{
clean_url(url)
}
\arguments{
\item{url}{The URL to clean}
}
\value{
A cleaned and formatted URL
}
\description{
Clean and format URL
}
\keyword{internal}
